SD & SSD [US012] Changed telephone number from int to long, to avoid int overflows Changes to ImportController methods

+++ src/main/java/pt/ipp/isep/dei/esoft/project/application/controller/RegisterEmployeeController.java
@@ -76,7 +76,7 @@ public class RegisterEmployeeController {
      * @param roles
      * @return
      */
-    public boolean registerEmployee(String name, String email, int passportCardNumber, int taxNumber, int telephoneNumber, Address address, int agencyID, ArrayList<Role> roles) {
+    public boolean registerEmployee(String name, String email, int passportCardNumber, int taxNumber, long telephoneNumber, Address address, int agencyID, ArrayList<Role> roles) {
         Employee employee = new Employee(name, email, passportCardNumber, taxNumber, telephoneNumber, address, roles, agencyID);
         return employeeRepository.addEmployee(employee);
     }
